requirements.txt
openai
python-dotenv
python-docx
pandas
openpyxl
tqdm
